single button to connect to hotspot
this commit has some major design changes imo. first button from left on appbar is for auto connect as wifi direct as always. second button is now the single click button for connecting to a hotspot(to become a bridge GM). third button onclick will bring up an alert dialog box, if group is created, it will display the passphrase of the group, if not, theres an option below to create a wifi direct group manually which will help us in debugging and presentation. \n plus some changes in xml file - connectToHotspot and discoverPeersOnly are interchanged, and discoverPeersOnly doesnt actually do what it says. will refactor it in the next commit