ledger working
ledger works with multiple devices. only drawback right now is that GO doesn't send it's ledger details to GOs directly. but when a GO or GM inputs a ledger, or connects to the network, the details are broadcasted to every other GM in the network. Only the existing GO records are not sent to newly connected GM, yet.